# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/contests/{contestId}/awards/generate:
    post:
      tags:
      - award-controller
      operationId: generate
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AwardGenerationResponse"
...
  /api/public/presentations/{contestId}:
    get:
      tags:
      - presentation-controller
      operationId: getPublished
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: mode
        in: query
        required: true
        schema:
          type: string
          enum:
          - SCREEN
          - LIVE
      - name: token
        in: query
        required: false
        schema:
          type: string
      - name: X-Broadcast-Token
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PresentationResponse"
...
  /api/admin/contests/{contestId}/exports/submissions.csv:
    get:
      tags:
      - data-export-controller
      operationId: submissions
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
  /api/admin/contests/{contestId}/exports/scoreboard.csv:
    get:
      tags:
      - data-export-controller
      operationId: scoreboard
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    PresentationConfigResponse:
      type: object
      properties:
        contestId:
          type: integer
          format: int64
        mode:
          type: string
          enum:
          - SCREEN
          - LIVE
        enabled:
          type: boolean
        title:
          type: string
        subtitle:
          type: string
        accentColor:
          type: string
        rowLimit:
          type: integer
          format: int32
        showAnnouncements:
          type: boolean
        announcementIntervalSeconds:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
...
    AwardGenerationResponse:
      type: object
      properties:
        categoriesProcessed:
          type: integer
          format: int32
        categoriesSkipped:
          type: integer
          format: int32
...
    FirstBloodEntry:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        problemAlias:
          type: string
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        submittedAt:
          type: string
          format: date-time
    PresentationAnnouncementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        body:
          type: string
        pinned:
          type: boolean
        publishedAt:
          type: string
          format: date-time
    PresentationResponse:
      type: object
      properties:
        contestId:
          type: integer
          format: int64
        contestName:
          type: string
        contestStatus:
          type: string
          enum:
          - DRAFT
          - FROZEN_CONFIG
          - RUNNING
          - PAUSED
          - ENDED
          - ARCHIVED
        startAt:
          type: string
          format: date-time
        freezeAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        serverTime:
          type: string
          format: date-time
        config:
          $ref: "#/components/schemas/PresentationConfigResponse"
        scoreboard:
          $ref: "#/components/schemas/ScoreboardResponse"
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/PresentationAnnouncementResponse"
    ScoreboardCell:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        alias:
          type: string
        solved:
          type: boolean
        attempts:
          type: integer
          format: int32
        wrongAttempts:
          type: integer
          format: int32
        firstAcceptedAt:
          type: string
          format: date-time
        penaltyMinutes:
          type: integer
          format: int64
        firstBlood:
          type: boolean
    ScoreboardProblem:
      type: object
      properties:
        problemId:
          type: integer
          format: int64
        alias:
          type: string
        name:
          type: string
    ScoreboardResponse:
      type: object
      properties:
        contestId:
          type: integer
          format: int64
        contestName:
          type: string
        variant:
          type: string
        groupName:
          type: string
        frozen:
          type: boolean
        freezeAt:
          type: string
          format: date-time
        generatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardProblem"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardRow"
        firstBlood:
          type: array
          items:
            $ref: "#/components/schemas/FirstBloodEntry"
    ScoreboardRow:
      type: object
      properties:
        rank:
          type: integer
          format: int32
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        school:
          type: string
        star:
          type: boolean
        participationType:
          type: string
        groupName:
          type: string
        solved:
          type: integer
          format: int32
        penaltyMinutes:
          type: integer
          format: int64
        latestAcceptedAt:
          type: string
          format: date-time
        cells:
          type: array
          items:
            $ref: "#/components/schemas/ScoreboardCell"
...